refactor: inline StoreBackend trait into store/mod.rs
Remove the single-trait backend.rs file and place the trait
definition directly in store/mod.rs.
